Misc: Used `assert_eq` and `assert_ne` and other code improvements
These are found by IDEA linter, which are good to have.

+++ kernel/src/process/syscalls/mod.rs
@@ -294,7 +294,7 @@ fn sys_read(all_state: &mut InterruptAllSavedState) -> SyscallResult {
     } else {
         bytes_read
     };
-    SyscallResult::Ok(bytes_read as u64)
+    SyscallResult::Ok(bytes_read)
 }
 
 fn sys_close(all_state: &mut InterruptAllSavedState) -> SyscallResult {
